Tested input_summary_table.
Added tests for functions:

- input_summary_row_ui, and
- input_summary_row_server.

The tests contain the simplest comprehensive checks I could get to work
at the moment, "comprehensive" meaning "covering the greatest possible
part of the code".

These tests check the output HTML code from the first function. They do
not check particular properties of the generated code for now, but they
should cover most of the code in the tested functions. They can be used
as a basis for expanding the checks or making them more specific.

Possibilities include counting observers before and after changing the
number of input files (the observers are already counted in the second
test).

Provided tests generate one warning. The warning is:

> (@test-input_summary_table.R#17) - argument is not an atomic vector;
> coercing

The text of the warning can be previewed by selecting a custom test
reporter in testthat.R. To do that, change:

  test_check("iaoreader")

to:

  test_check("iaoreader", reporter="summary")

A recommended requirement on the "stringi" package in the DESCRIPTION
file has been added to fix a CI error in AppVeyor.